Mise en place des DTO, services CRUD et logique métier, exceptions et controllers

+++ ODM_Backend/src/main/java/org/example/odm_backend/services/serviceImpl/EquipeServiceImpl.java
@@ -1,11 +1,18 @@
 package org.example.odm_backend.services.serviceImpl;
 
+import org.example.odm_backend.dtos.EquipeDto.EquipeFilterDTO;
 import org.example.odm_backend.dtos.EquipeDto.EquipeRequestDTO;
 import org.example.odm_backend.dtos.EquipeDto.EquipeResponseDTO;
 import org.example.odm_backend.entities.Equipe;
+import org.example.odm_backend.entities.Projet;
+import org.example.odm_backend.exceptions.DuplicateResourceException;
+import org.example.odm_backend.exceptions.NotFoundException;
 import org.example.odm_backend.mappers.EquipeMapper;
 import org.example.odm_backend.repositories.EquipeRepository;
+import org.example.odm_backend.repositories.ProjetRepository;
 import org.example.odm_backend.services.serviceInterface.EquipeService;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,29 +21,36 @@ import java.util.List;
 public class EquipeServiceImpl implements EquipeService {
 
     private final EquipeRepository equipeRepository;
+    private final ProjetRepository projetRepository;
     private final EquipeMapper equipeMapper;
 
-    public EquipeServiceImpl(EquipeRepository equipeRepository, EquipeMapper equipeMapper){
+    public EquipeServiceImpl(
+            EquipeRepository equipeRepository,
+            ProjetRepository projetRepository,
+            EquipeMapper equipeMapper
+    ) {
         this.equipeRepository = equipeRepository;
+        this.projetRepository = projetRepository;
         this.equipeMapper = equipeMapper;
     }
 
     @Override
     public EquipeResponseDTO addEquipe(EquipeRequestDTO dto) {
-        Equipe equipe = new Equipe();
-        equipe.setNomEquipe(dto.nomEquipe());
-        equipe = equipeRepository.save(equipe);
 
-        return equipeMapper.toResponse(equipe);
-    }
+        if (equipeRepository.existsByNomEquipe(dto.nomEquipe())) {
+            throw new DuplicateResourceException("Cette équipe existe déjà");
+        }
 
-    @Override
-    public EquipeResponseDTO updateEquipe(Long id, EquipeRequestDTO dto) {
-        Equipe equipe = equipeRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Equipe pas trouvée"));
+        Equipe equipe = equipeMapper.toEntity(dto);
 
-        if (dto.nomEquipe() != null) {
-            equipe.setNomEquipe(dto.nomEquipe());
+        if (dto.projetIds() != null && !dto.projetIds().isEmpty()) {
+            List<Projet> projets = projetRepository.findAllById(dto.projetIds());
+
+            if (projets.size() != dto.projetIds().size()) {
+                throw new NotFoundException("Un ou plusieurs projets sont introuvables");
+            }
+
+            equipe.setProjets(projets);
         }
 
         equipe = equipeRepository.save(equipe);
@@ -45,22 +59,56 @@ public class EquipeServiceImpl implements EquipeService {
     }
 
     @Override
+    public EquipeResponseDTO updateEquipe(Long id, EquipeRequestDTO dto) {
+
+        Equipe equipe = equipeRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Equipe non trouvée"));
+
+        if (dto.nomEquipe() != null &&
+                !dto.nomEquipe().isBlank() &&
+                !dto.nomEquipe().equals(equipe.getNomEquipe())) {
+
+            if (equipeRepository.existsByNomEquipe(dto.nomEquipe())) {
+                throw new DuplicateResourceException("Cette équipe existe déjà");
+            }
+
+            equipe.setNomEquipe(dto.nomEquipe());
+        }
+
+        if (dto.projetIds() != null) {
+            List<Projet> projets = projetRepository.findAllById(dto.projetIds());
+            equipe.setProjets(projets);
+        }
+
+        equipeRepository.save(equipe);
+
+        return equipeMapper.toResponse(equipe);
+    }
+
+    @Override
     public void deleteEquipe(Long id) {
+
+        if (!equipeRepository.existsById(id)) {
+            throw new NotFoundException("Equipe non trouvée");
+        }
+
         equipeRepository.deleteById(id);
     }
 
     @Override
     public EquipeResponseDTO getById(Long id) {
+
         Equipe equipe = equipeRepository.findById(id)
-                .orElseThrow(()-> new RuntimeException("Equipe pas trouvée"));
-        return null;
+                .orElseThrow(() -> new NotFoundException("Equipe non trouvée"));
+
+        return equipeMapper.toResponse(equipe);
     }
 
     @Override
-    public List<EquipeResponseDTO> getAll() {
-        return equipeRepository.findAll()
-                .stream()
-                .map(equipeMapper::toResponse)
-                .toList();
+    public Page<EquipeResponseDTO> search(EquipeFilterDTO filter, Pageable pageable) {
+        return equipeRepository.search(
+                filter.nomEquipe(),
+                pageable
+        ).map(equipeMapper::toResponse);
     }
 }
